move top buttons to VerticalFeed
